+++ src/main/kotlin/racingcar/Car.kt
@@ -1,4 +1,4 @@
-package racinggame
+package racingcar
 
 class Car(
     var position: Int = 0
