+++ src/test/kotlin/racingcar/CarTest.kt
@@ -1,4 +1,4 @@
-package racinggame
+package racingcar
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
